python-analyzer/requirements.txt
# ============================================
# ETH Crypto Prediction System - Python Dependencies
# Compatible with Ubuntu 22.04 + Python 3.10/3.11/3.12
# ============================================

# === 数据处理与科学计算 ===
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0

# === 技术分析 ===
ta==0.11.0
ta-lib==0.4.28          # 需要先安装系统库: sudo apt install libta-lib-dev
pandas-ta==0.3.14b1

# === 机器学习 ===
scikit-learn==1.4.1
xgboost==2.0.3
lightgbm==4.3.0

# === 深度学习 ===
torch==2.2.1
tensorflow==2.15.0

# === 强化学习 ===
stable-baselines3==2.2.1
gymnasium==0.29.1

# === 数据可视化 ===
matplotlib==3.8.3
plotly==5.19.0
dash==2.16.1
dash-bootstrap-components==1.5.0
mplfinance==0.12.10b0
kaleido==0.2.1

# === API与网络请求 ===
requests==2.31.0
websocket-client==1.7.0
aiohttp==3.9.3
ccxt==4.2.70           # 统一加密货币交易所API

# === 数据库 ===
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# === 工具库 ===
python-dotenv==1.0.1
schedule==1.2.1
joblib==1.3.2
tqdm==4.66.2
colorama==0.4.6

# === 情绪分析 (NLP) ===
transformers==4.38.2
textblob==0.18.0

# === Web框架 (API) ===
flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
